New turbo decoder implementation, minor updates here and there

+++ trunk/openair1/PHY/CODING/lte_interleaver_inline.h
@@ -1,4 +1,4 @@
-extern unsigned short threegpplte_interleaver_output;
+extern unsigned int threegpplte_interleaver_output;
 extern unsigned int threegpplte_interleaver_tmp;
 
 extern inline void threegpplte_interleaver_reset(void) {
